// src/enums/common.ts
export enum Difficulty {
    BEGINNER = 'Beginner',
    EASY = 'Easy',
    MEDIUM = 'Medium',
    HARD = 'Hard',
    PROFESSIONAL = 'Professional',
    EXPERT = 'Expert'
}

export enum LearnerLevel {
    PRE_KG = 'Pre-KG',
    SCHOOL = 'School',
    COLLEGE = 'College',
    PROFESSIONAL = 'Professional',
    PHD = 'PhD'
}